+++ extras/tests/JsonDocument/subscript.cpp
@@ -22,6 +22,12 @@ TEST_CASE("JsonDocument::operator[]") {
       REQUIRE(cdoc[std::string("hello")] == "world");
     }
 
+    SECTION("JsonVariant") {
+      doc["key"] = "hello";
+      REQUIRE(doc[doc["key"]] == "world");
+      REQUIRE(cdoc[cdoc["key"]] == "world");
+    }
+
     SECTION("supports operator|") {
       REQUIRE((doc["hello"] | "nope") == std::string("world"));
       REQUIRE((doc["world"] | "nope") == std::string("nope"));
@@ -31,8 +37,16 @@ TEST_CASE("JsonDocument::operator[]") {
   SECTION("array") {
     deserializeJson(doc, "[\"hello\",\"world\"]");
 
-    REQUIRE(doc[1] == "world");
-    REQUIRE(cdoc[1] == "world");
+    SECTION("int") {
+      REQUIRE(doc[1] == "world");
+      REQUIRE(cdoc[1] == "world");
+    }
+
+    SECTION("JsonVariant") {
+      doc[2] = 1;
+      REQUIRE(doc[doc[2]] == "world");
+      REQUIRE(cdoc[doc[2]] == "world");
+    }
   }
 }